Prevent a fingerprinting issue with various login pages
... by not handling the requests as image requests
or fast-redirecting them.

Without the added section a request to a blocked or
redirected login URL could be misdetected by third
parties as the user being logged in to the given site,
thus making fingerprinting Privoxy users easier.

Note that this does not prevent the fingerprinting issue
if the client is actually logged in. For details see:
https://robinlinus.github.io/socialmedia-leak/

Doing that would probably be too invasive for a default
configuration.